requirement.txt
faiss-cpu 
neo4j
zhipuai
numpy
pandas
jieba
pandas
py2neo
python-dotenv